+++ src/Index.tsx
@@ -1,6 +1,4 @@
-import 'babel-polyfill';
-// Promise.finally isn't in the base polyfill
-import 'core-js/fn/promise/finally';
+import '@babel/polyfill';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { module as angularModule, bootstrap } from 'angular';
@@ -24,6 +22,7 @@ import { lazyInjector } from './lazyInjector';
 import { safariTouchFix } from './safari-touch-fix';
 import Root from './Root';
 import updateCSSVariables from './app/css-variables';
+import setupRateLimiter from './app/bungie-api/rate-limit-config';
 
 polyfill({
   holdToDrag: 300,
@@ -41,6 +40,7 @@ initi18n().then(() => {
     'ngInject';
     lazyInjector.$injector = $injector;
     updateCSSVariables();
+    setupRateLimiter();
 
     ReactDOM.render(<Root />, document.getElementById('app'));
   });