+++ src/app/d2-loadout-builder/generated-sets/GeneratedSetItem.tsx
@@ -1,21 +1,24 @@
 import * as React from 'react';
-import { D2Item } from '../../inventory/item-types';
-import { filterPlugs } from './utils';
-import PressTip from '../../dim-ui/PressTip';
-import PlugTooltip from './PlugTooltip';
 import BungieImage from '../../dim-ui/BungieImage';
-import ConnectedInventoryItem from '../../inventory/ConnectedInventoryItem';
-import DraggableInventoryItem from '../../inventory/DraggableInventoryItem';
-import ItemPopupTrigger from '../../inventory/ItemPopupTrigger';
+import PressTip from '../../dim-ui/PressTip';
+import { D2Item } from '../../inventory/item-types';
+import LoadoutBuilderItem from '../LoadoutBuilderItem';
+import { LockedItemType } from '../types';
+import PlugTooltip from './PlugTooltip';
+import { filterPlugs } from './utils';
 
-export default function GeneratedSetItem({ item }: { item: D2Item }) {
+export default function GeneratedSetItem({
+  item,
+  locked,
+  onExclude
+}: {
+  item: D2Item;
+  locked: LockedItemType[];
+  onExclude(item: LockedItemType): void;
+}) {
   return (
     <div className="generated-build-items">
-      <DraggableInventoryItem item={item}>
-        <ItemPopupTrigger item={item}>
-          <ConnectedInventoryItem item={item} />
-        </ItemPopupTrigger>
-      </DraggableInventoryItem>
+      <LoadoutBuilderItem item={item} locked={locked} onExclude={onExclude} />
       {item!.sockets &&
         item!.sockets!.categories.length === 2 &&
         // TODO: look at plugs that we filtered on to see if they match selected perk or not.