+++ src/app/d2-loadout-builder/generated-sets/GeneratedSets.tsx
@@ -1,24 +1,26 @@
 import { t } from 'i18next';
 import * as React from 'react';
-import * as _ from 'lodash';
-import { Loadout, dimLoadoutService } from '../../loadout/loadout.service';
-import { ArmorSet, LockType } from '../types';
-import { getSetsForTier, getSetTiers } from './utils';
-import GeneratedSetButtons from './GeneratedSetButtons';
+import { InventoryBucket } from '../../inventory/inventory-buckets';
 import { DimStore } from '../../inventory/store-types';
+import { dimLoadoutService, Loadout } from '../../loadout/loadout.service';
 import LoadoutDrawer from '../../loadout/LoadoutDrawer';
+import { ArmorSet, LockedItemType } from '../types';
+import GeneratedSetButtons from './GeneratedSetButtons';
 import GeneratedSetItem from './GeneratedSetItem';
+import { getSetsForTier, getSetTiers, isD2Item, toggleLockedItem } from './utils';
 
 interface Props {
   processRunning: number;
   selectedStore?: DimStore;
   processedSets: ArmorSet[];
-  lockedMap: { [bucketHash: number]: LockType };
+  lockedMap: { [bucketHash: number]: LockedItemType[] };
+  onLockChanged(bucket: InventoryBucket, locked?: LockedItemType[]): void;
 }
 
 interface State {
   minimumPower: number;
   selectedTier: string;
+  shownSets: number;
 }
 
 let matchedSets: ArmorSet[] = [];
@@ -29,7 +31,8 @@ let matchedSets: ArmorSet[] = [];
 export default class GeneratedSets extends React.Component<Props, State> {
   state: State = {
     selectedTier: '7/7/7',
-    minimumPower: 0
+    minimumPower: 0,
+    shownSets: 10
   };
 
   // Set the loadout property to show/hide the loadout menu
@@ -39,13 +42,38 @@ export default class GeneratedSets extends React.Component<Props, State> {
 
   componentWillReceiveProps(props: Props) {
     if (props.processedSets !== this.props.processedSets) {
-      this.setState({ minimumPower: 0 });
+      this.setState({ minimumPower: 0, shownSets: 10 });
     }
   }
 
+  showMore = () => {
+    this.setState({ shownSets: this.state.shownSets + 10 });
+  };
+
+  setSelectedTier = (element) => {
+    this.setState({ shownSets: 10, selectedTier: element.target.value });
+  };
+
+  setMinimumPower = (element) => {
+    this.setState({ shownSets: 10, minimumPower: parseInt(element.target.value, 10) });
+  };
+
+  toggleLockedItem = (lockedItem: LockedItemType) => {
+    if (!isD2Item(lockedItem.item)) {
+      return;
+    }
+    const bucket = lockedItem.item.bucket;
+    toggleLockedItem(
+      lockedItem,
+      bucket,
+      this.props.onLockChanged,
+      this.props.lockedMap[bucket.hash]
+    );
+  };
+
   render() {
     const { processRunning, lockedMap, selectedStore } = this.props;
-    const { selectedTier, minimumPower } = this.state;
+    const { selectedTier, minimumPower, shownSets } = this.state;
 
     if (processRunning > 0) {
       return <h3>{t('LoadoutBuilder.Loading', { loading: processRunning })}</h3>;
@@ -77,12 +105,7 @@ export default class GeneratedSets extends React.Component<Props, State> {
     return (
       <>
         <h3>{t('LoadoutBuilder.SelectPower')}</h3>
-        <select
-          value={minimumPower}
-          onChange={(element) => {
-            this.setState({ minimumPower: parseInt(element.target.value, 10) });
-          }}
-        >
+        <select value={minimumPower} onChange={this.setMinimumPower}>
           {powerLevelOptions.map((power) => {
             if (power === 0) {
               return (
@@ -96,12 +119,7 @@ export default class GeneratedSets extends React.Component<Props, State> {
         </select>
 
         <h3>{t('LoadoutBuilder.SelectTier')}</h3>
-        <select
-          value={currentTier}
-          onChange={(element) => {
-            this.setState({ selectedTier: element.target.value });
-          }}
-        >
+        <select value={currentTier} onChange={this.setSelectedTier}>
           {setTiers.map((tier) => (
             <option key={tier} disabled={tier.charAt(0) === '-'}>
               {tier}
@@ -110,8 +128,8 @@ export default class GeneratedSets extends React.Component<Props, State> {
         </select>
 
         <h3>{t('LoadoutBuilder.GeneratedBuilds')}</h3>
-        {matchedSets.map((set, index) => (
-          <div className="generated-build" key={index}>
+        {matchedSets.slice(0, shownSets).map((set) => (
+          <div className="generated-build" key={set.id}>
             <GeneratedSetButtons
               set={set}
               store={selectedStore!}
@@ -119,11 +137,21 @@ export default class GeneratedSets extends React.Component<Props, State> {
             />
             <div className="sub-bucket">
               {Object.values(set.armor).map((item) => (
-                <GeneratedSetItem key={item.index} item={item} />
+                <GeneratedSetItem
+                  key={item.index}
+                  item={item}
+                  locked={lockedMap[item.bucket.hash]}
+                  onExclude={this.toggleLockedItem}
+                />
               ))}
             </div>
           </div>
         ))}
+        {matchedSets.length > shownSets && (
+          <button className="dim-button" onClick={this.showMore}>
+            {t('LoadoutBuilder.ShowMore')}
+          </button>
+        )}
         <LoadoutDrawer />
       </>
     );