+++ src/app/d2-loadout-builder/generated-sets/utils.ts
@@ -1,6 +1,9 @@
+import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import { t } from 'i18next';
 import * as _ from 'lodash';
-import { ArmorSet, LockType } from '../types';
+import { InventoryBucket } from '../../inventory/inventory-buckets';
+import { D2Item } from '../../inventory/item-types';
+import { ArmorSet, LockedItemType } from '../types';
 
 /**
  *  Filter out plugs that we don't want to show in the perk dropdown.
@@ -24,59 +27,46 @@ export function filterPlugs(socket) {
  */
 export function getSetsForTier(
   setMap: ArmorSet[],
-  lockedMap: { [bucketHash: number]: LockType },
+  lockedMap: { [bucketHash: number]: LockedItemType[] },
   tier: string
 ): ArmorSet[] {
   const matchedSets: ArmorSet[] = [];
-  let count = 0;
 
   Object.values(setMap).forEach((set) => {
-    if (count > 10) {
-      return;
-    }
-
     if (set.tiers.includes(tier)) {
       matchedSets.push(set);
-      count++;
     }
   });
 
   // Sort based on power level
-  matchedSets.sort((a, b) => b.power - a.power);
+  let sortedSets = _.sortBy(matchedSets, (set) => -set.power);
 
   // Prioritize list based on number of matched perks
   Object.keys(lockedMap).forEach((bucket) => {
-    // if there are locked items for this bucket
-    if (lockedMap[bucket] && lockedMap[bucket].items.length && lockedMap[bucket].type === 'perk') {
-      // Sort based on what sets have the most matched perks
-      matchedSets.sort((a, b) => {
-        return (
-          _.sumBy(b.armor, (item) => {
-            if (!item || !item.sockets) {
-              return 0;
-            }
-            return item.sockets.sockets.filter((slot) =>
-              slot.plugOptions.some((perk) =>
-                lockedMap[bucket].items.find((lockedPerk) => lockedPerk.hash === perk.plugItem.hash)
-              )
-            ).length;
-          }) -
-          _.sumBy(a.armor, (item) => {
-            if (!item || !item.sockets) {
-              return 0;
-            }
-            return item.sockets.sockets.filter((slot) =>
-              slot.plugOptions.some((perk) =>
-                lockedMap[bucket].items.find((lockedPerk) => lockedPerk.hash === perk.plugItem.hash)
-              )
-            ).length;
-          })
-        );
-      });
+    // if there are locked perks for this bucket
+    if (lockedMap[bucket] === undefined) {
+      return;
     }
+    const lockedPerks = lockedMap[bucket].filter((lockedItem) => lockedItem.type === 'perk');
+    if (!lockedPerks.length) {
+      return;
+    }
+    // Sort based on what sets have the most matched perks
+    sortedSets = _.sortBy(sortedSets, (set) => {
+      return -_.sumBy(set.armor, (item) => {
+        if (!item || !item.sockets) {
+          return 0;
+        }
+        return item.sockets.sockets.filter((slot) =>
+          slot.plugOptions.some((perk) =>
+            lockedPerks.find((lockedPerk) => lockedPerk.item.hash === perk.plugItem.hash)
+          )
+        ).length;
+      });
+    });
   });
 
-  return matchedSets;
+  return sortedSets;
 }
 
 /**
@@ -111,4 +101,35 @@ export function getSetTiers(armorSets: ArmorSet[]): string[] {
   }
 
   return setTiers;
+}
+
+export function toggleLockedItem(
+  lockedItem: LockedItemType,
+  bucket: InventoryBucket,
+  onLockChanged: (bucket: InventoryBucket, locked?: LockedItemType[]) => void,
+  locked?: LockedItemType[]
+) {
+  let newLockedItems: LockedItemType[] = [];
+  if (locked && locked[0].type !== 'item') {
+    newLockedItems = Array.from(locked);
+  }
+
+  const existingIndex = newLockedItems.findIndex((existing) => existing.item === lockedItem.item);
+  if (existingIndex > -1) {
+    newLockedItems.splice(existingIndex, 1);
+  } else {
+    newLockedItems.push(lockedItem);
+  }
+
+  onLockChanged(bucket, newLockedItems.length === 0 ? undefined : newLockedItems);
+}
+
+export function isInventoryItemDefinition(
+  item: D2Item | DestinyInventoryItemDefinition
+): item is DestinyInventoryItemDefinition {
+  return Boolean((item as DestinyInventoryItemDefinition).displayProperties);
+}
+
+export function isD2Item(item: D2Item | DestinyInventoryItemDefinition): item is D2Item {
+  return Boolean((item as D2Item).name);
 }