+++ src/app/d2-loadout-builder/locked-armor/LockedArmor.tsx
@@ -3,18 +3,19 @@ import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import * as React from 'react';
 import { InventoryBucket } from '../../inventory/inventory-buckets';
 import { D2Item } from '../../inventory/item-types';
-import { LockType } from '../types';
+import { LockedItemType } from '../types';
 import LoadoutBucketDropTarget from './LoadoutBucketDropTarget';
 import './lockedarmor.scss';
 import LockedItem from './LockedItem';
 import LockablePopup from './popup/LockablePopup';
+import { toggleLockedItem } from '../generated-sets/utils';
 
 interface Props {
   bucket: InventoryBucket;
   items: { [itemHash: number]: D2Item[] };
   perks: Set<DestinyInventoryItemDefinition>;
-  locked?: LockType;
-  onLockChanged(bucket: InventoryBucket, locked?: LockType): void;
+  locked?: LockedItemType[];
+  onLockChanged(bucket: InventoryBucket, locked?: LockedItemType[]): void;
 }
 
 interface State {
@@ -34,11 +35,17 @@ export default class LockedArmor extends React.Component<Props & UIViewInjectedP
     this.setState({ isOpen: false });
   };
 
-  setLockedItem = (lockedItem: D2Item) => {
-    this.props.onLockChanged(this.props.bucket, {
-      type: 'item',
-      items: [lockedItem]
-    });
+  setLockedItem = (item: D2Item) => {
+    this.props.onLockChanged(this.props.bucket, [
+      {
+        type: 'item',
+        item
+      }
+    ]);
+  };
+
+  toggleLockedItem = (lockedItem: LockedItemType) => {
+    toggleLockedItem(lockedItem, this.props.bucket, this.props.onLockChanged, this.props.locked);
   };
 
   reset = () => {
@@ -52,10 +59,13 @@ export default class LockedArmor extends React.Component<Props & UIViewInjectedP
     return (
       <div className="locked-item">
         <LoadoutBucketDropTarget bucketType={bucket.type!} onItemLocked={this.setLockedItem}>
-          {locked &&
-            locked.items &&
-            locked.items.length !== 0 && <div className="close" onClick={this.reset} />}
-          <LockedItem {...{ locked, bucket, toggleOpen: this.openPerkSelect }} />
+          {locked && locked.length !== 0 && <div className="close" onClick={this.reset} />}
+          <LockedItem
+            locked={locked}
+            bucket={bucket}
+            toggleOpen={this.openPerkSelect}
+            onExclude={this.toggleLockedItem}
+          />
         </LoadoutBucketDropTarget>
         <LockablePopup
           bucket={bucket}