+++ src/app/d2-loadout-builder/locked-armor/popup/SelectableBungieImage.tsx
@@ -2,6 +2,7 @@ import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import * as classNames from 'classnames';
 import * as React from 'react';
 import BungieImage from '../../../dim-ui/BungieImage';
+import { LockedItemType } from '../../types';
 
 export default function SelectableBungieImage({
   perk,
@@ -11,11 +12,11 @@ export default function SelectableBungieImage({
 }: {
   perk: DestinyInventoryItemDefinition;
   selected: boolean;
-  onLockedPerk(perk: DestinyInventoryItemDefinition): void;
+  onLockedPerk(perk: LockedItemType): void;
   onHoveredPerk(perk: {}): void;
 }) {
   const handleClick = () => {
-    onLockedPerk(perk);
+    onLockedPerk({ type: 'perk', item: perk });
   };
   const handleHover = () => {
     onHoveredPerk(perk.displayProperties);