+++ src/app/inventory/ClearNewItems.tsx
@@ -9,6 +9,7 @@ import { NewItemsService } from './store/new-items.service';
 import './ClearNewItems.scss';
 import { connect } from 'react-redux';
 import { RootState } from '../store/reducers';
+import { AppIcon, thumbsUpIcon } from '../shell/icons';
 
 interface ProvidedProps {
   account: DestinyAccount;
@@ -57,7 +58,7 @@ class ClearNewItems extends React.Component<Props> {
     return (
       <div className="clear-new-items">
         <button onClick={this.clearNewItems} title={t('Hotkey.ClearNewItemsTitle')}>
-          <i className="fa fa-thumbs-up" /> <span>{t('Hotkey.ClearNewItems')}</span>
+          <AppIcon icon={thumbsUpIcon} /> <span>{t('Hotkey.ClearNewItems')}</span>
         </button>
       </div>
     );