+++ src/app/shell/icons/ReviewIcon.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { dtrRatingColor } from '../dimAngularFilters.filter';
+import AppIcon from './AppIcon';
+import { starIcon } from './Library';
+
+export const RatingIcon = ({ rating }: { rating: number }) => (
+  <AppIcon icon={starIcon} style={dtrRatingColor(rating)} />
+);