+++ src/app/vendors/vendor-items.component.ts
@@ -1,6 +1,6 @@
 import { settings } from '../settings/settings';
 import template from './vendor-items.html';
-import { IComponentOptions, IController } from 'angular';
+import { IComponentOptions, IController, IScope } from 'angular';
 import store from '../store/store';
 import { toggleCollapsedSection } from '../settings/actions';
 
@@ -19,13 +19,19 @@ export const VendorItems: IComponentOptions = {
 function VendorItemsCtrl(
   this: IController & {
     settings: typeof settings;
-  }
+  },
+  $scope: IScope
 ) {
   'ngInject';
 
   const vm = this;
 
-  vm.settings = settings;
+  $scope.$watch(
+    () => settings,
+    () => {
+      vm.settings = settings;
+    }
+  );
 
   vm.toggleSection = (id) => {
     store.dispatch(toggleCollapsedSection(id));